pyproject.toml
[tool.poetry]
name = "workana-scraping"
version = "0.1.0"
description = ""
authors = ["Luiz Paulo Coutinho <coutinholps@gmail.com>"]
readme = "README.md"
packages = [{include = "workana_scraping"}]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.9.1"
pandas = "^2.0.1"
webdriver-manager = "^3.8.6"
pytelegrambotapi = "^4.11.0"
python-dotenv = "^1.0.0"
undetected-chromedriver = "^3.4.7"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = {extras = ["jupyter"], version = "^23.3.0"}
ipykernel = "^6.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
